+++ src/main/java/api/buyhood/domain/product/entity/Product.java
@@ -2,15 +2,7 @@ package api.buyhood.domain.product.entity;
 
 import api.buyhood.domain.store.entity.Store;
 import api.buyhood.entity.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
